% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{bp_batch}
\alias{bp_batch}
\title{Batch Process a File or Data Frame}
\usage{
bp_batch(
  x,
  columns = c(idcol = 1, age = 2, male = 3, height = 4, sbp = 5, dbp = 6),
  ...
)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{columns}{named integer vector for identifying the id, age, sex, height,
sbp and dpb columns}

\item{...}{no currently used}
}
\value{
a \code{data.frame} with the \code{columns} from \code{x} with two
additional columns: \code{sbp_percentile} and \code{dbp_percentile}.
}
\description{
Given a \code{data.frame} of age, sex, height, and blood pressures, return
the \code{data.frame} with two additional columns reporting the percentiles
for the systolic and diastolic blood pressures.
}
\examples{

d <- read.csv(system.file("example_data", "for_batch.csv", package = "pedbp"))
d
bp_batch(d)

# use with a data.frame with columns in a different order
d2 <- data.frame(
                   age = d$age_months
                 , x1 = runif(5)
                 , sbp = d$sbp..mmHg.
                 , ht  = d$height..cm.
                 , sex = d$male
                 , dbp = d$dbp..mmHg.
                 , id  = d$pid
                 )
d2
bp_batch(d2, columns = c(idcol = 7, age = 1, male = 5, height = 4, sbp = 3, dbp = 6))

}
