% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bp_chart}
\alias{bp_chart}
\title{Blood Pressure Chart}
\usage{
bp_chart(height, height_unit, age, male = NULL, sex = NULL, ...)
}
\arguments{
\item{height}{numeric value for the height of the subject}

\item{height_unit}{character string indicating if the \code{height} value is
in inches or centimeters.  Acceptable inputs are form the set \code{c("in",
"inches", "cm", "centimeters")}.}

\item{age}{In years, expected integer value, will round to nearest integer
value if non-integer value is passed in.}

\item{male}{an integer value (1 = male, 0 = female)}

\item{sex}{an alternative to using the \code{male} argument, expecting a
character string of length one within the case insensitive set, \code{c("m",
"male", "f", "female")}.}

\item{...}{Pass through}
}
\description{
Produce a graphic showing the regions of blood pressure percentiles by age
(in years) and height (percentile).
}
