% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_distribution.R
\name{bp_distribution}
\alias{bp_distribution}
\alias{p_bp}
\alias{q_bp}
\title{Estimate Pediatric Blood Pressure Distribution}
\usage{
p_bp(q_sbp, q_dbp, age, male, height = NA, height_percentile = 0.5, ...)

q_bp(p_sbp, p_dbp, age, male, height = NA, height_percentile = 0.5, ...)
}
\arguments{
\item{q_sbp}{a vector of systolic blood pressures}

\item{q_dbp}{a vector of diastolic blood pressures}

\item{age}{numeric age, in months}

\item{male}{integer value, 1 = male, 0 = female, indicating sex of the
patient}

\item{height}{numeric, in centimeters, can be missing.  This is the length
for patients under three years of age}

\item{height_percentile}{default height percentile to use if \code{height} is
missing.}

\item{...}{not currently used}

\item{p_sbp}{a vector of systolic blood percentiles}

\item{p_dbp}{a vector of diastolic blood percentiles}
}
\description{
Percentile and quantile functions for pediatric blood pressure.
}
\details{
The areguments \code{age}, \code{male}, and \code{height} are expected to all
have lenght 1.  This is to simplify the base code and use of a look up table.
If you have the needed to quickly get the percentiles for a collection of
patients consider using \link{\code{p_bp_batch}}.
}
\examples{

x <- p_bp( q_sbp = 100, q_dbp = 60, age = 8, male = 0)
x
str(x)

x <- p_bp(q_sbp = c(NA, 82), q_dbp = c(60, 72), age = 9.2, male = 0)
x
str(x)

x <- p_bp(q_sbp = c(NA, 82), q_dbp = c(60, 72), age = 29.2, male = 0, height = 82.8)
x
str(x)

x <- q_bp(p_sbp = 0.78, p_dbp = 0.65, age = 8, male = 0)
x
str(x)

}
\seealso{
\code{vignette("bp-distriution", package = "pedbp")},
\link{\code{p_bp_batch}},
\link{\code{q_bp_batch}}
}
