% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flynn2017.R
\docType{data}
\name{flynn2017}
\alias{flynn2017}
\title{Blood Pressure by Age and Height Defined by Flynn et.al.(2017)}
\format{
a data.frame
}
\usage{
flynn2017
}
\description{
A data set resulting from the transcription of Tables 4 and 5 from Flynn
et.al (2017)
}
\references{
Flynn JT, Kaelber DC, Baker-Smith CM, et al. Clinical Practice Guideline for
Screening and Management of High Blood Pressure in Children and Adolescents.
Pediatrics. 2017;140(3):e20171904
}
\keyword{datasets}
