% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bp.R
\name{get_bp}
\alias{get_bp}
\title{Estimate BP from percentile for Age, Sex and Height}
\usage{
get_bp(
  sbp_percentile = NA,
  dbp_percentile = NA,
  age,
  male,
  height = NA,
  height_percentile = 0.5,
  ...
)
}
\arguments{
\item{sbp_percentile}{numeric between 0, 1}

\item{dbp_percentile}{numeric between 0, 1}

\item{age}{age in months}

\item{male}{integer value, 1 = male, 0 = female}

\item{height}{numeric, in centimeters, can be missing.}

\item{height_percentile}{default height percentile to use if \code{height} is
missing.}

\item{...}{not currently used}
}
\description{
Estimate BP from percentile for Age, Sex and Height
}
\examples{
get_bp(sbp_percentile = 0.9, dbp = 0.60, age = 8, male = 0)
}
