% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bp_percentile.R
\name{get_bp_percentile}
\alias{get_bp_percentile}
\title{Estimate Pediatric Blood Pressure Percentiles}
\usage{
get_bp_percentile(
  sbp,
  dbp,
  age,
  male,
  height = NA,
  height_percentile = 0.5,
  ...
)
}
\arguments{
\item{sbp}{Systolic blood pressure, mmHg}

\item{dbp}{Systolic blood pressure, mmHg}

\item{age}{age in months}

\item{male}{integer value, 1 = male, 0 = female}

\item{height}{numeric, in centimeters, can be missing.}

\item{height_percentile}{default height percentile to use if \code{height} is
missing.}

\item{...}{not currently used}
}
\description{
Estimate the percentile for blood pressure and height for a pediatric patient
based on sex and age and height.  Sex and age are required, height is not.
}
\examples{
get_bp_percentile(sbp = 100, dbp = 60, age = 8, male = 0)
}
