% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_height_percentile.R
\name{get_height_percentile}
\alias{get_height_percentile}
\title{Approximate Height Percentiles}
\usage{
get_height_percentile(age, male, height, ...)
}
\arguments{
\item{age}{age in months}

\item{male}{integer value, 1 = male, 0 = female}

\item{height}{in centimeters}

\item{...}{not currently used.}
}
\description{
Based on CDC data, height percentiles are estimated by age and sex, and by
interpolating CDC published quantiles with a Guassian distribution.
}
\details{
The \code{\link{cdc_length_for_age}} data was interploated to derive a mean
and standard deviation for a Guassian approximation of the height
percentiles.
}
\examples{

 e <- new.env()
 utils::data(list = "cdc_length_for_age", package = "pedbp", envir = e)

 e$cdc_length_for_age[130, ]
 get_height_percentile(age = 13.5, male = 0, height = 73.533349)
 get_height_percentile(age = 13.5, male = 0, height = 77)

}
