% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height_percentile}
\alias{height_percentile}
\title{Height Percentile}
\usage{
height_percentile(height, height_unit, age, male = NULL, sex = NULL, ...)
}
\arguments{
\item{height}{numeric value for the height of the subject}

\item{height_unit}{character string indicating if the \code{height} value is
in inches or centimeters.  Acceptable inputs are form the set \code{c("in",
"inches", "cm", "centimeters")}.}

\item{age}{In years, expected integer value, will round to nearest integer
value if non-integer value is passed in.}

\item{male}{an integer value (1 = male, 0 = female)}

\item{sex}{an alternative to using the \code{male} argument, expecting a
character string of length one within the case insensitive set, \code{c("m",
"male", "f", "female")}.}

\item{...}{Pass through}
}
\description{
Given the age, in years, sex, and height, in inches or centimeters, of a
child or adolescent, return the nearest height percentile as defined by the
data provided in Flynn et.al. (2017).
}
\examples{

height_percentile(83.2, "cm", age = 12, sex = "M")

}
\references{
Flynn JT, Kaelber DC, Baker-Smith CM, et al. Clinical Practice Guideline for
Screening and Management of High Blood Pressure in Children and Adolescents.
Pediatrics. 2017;140(3):e20171904
}
