% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lms_bmi}
\alias{lms_bmi}
\title{LMS data and distribution estimates}
\usage{
lms_bmi()
}
\arguments{
\item{age}{(in months)}

\item{stature}{(in cm)}
}
\description{
LMS data and distribution estimates
}
