% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lo2013}
\alias{lo2013}
\title{Mean and Standard Deviation of Blood Pressure, Height, Weight, and BMI from Lo et.al. (2013)}
\format{
a data.frame
}
\usage{
lo2013
}
\description{
Data from Appendix tables 1 and 2.
}
\references{
Lo, Joan C., et al. "Prehypertension and hypertension in community-based
pediatric practice." Pediatrics 131.2 (2013): e415-e424.
}
\keyword{datasets}
