% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbp-defunt.R
\name{p_length_for_age_inf}
\alias{p_length_for_age_inf}
\alias{q_length_for_age_inf}
\alias{z_length_for_age_inf}
\alias{p_weight_for_age_inf}
\alias{q_weight_for_age_inf}
\alias{z_weight_for_age_inf}
\alias{p_weight_for_length_inf}
\alias{q_weight_for_length_inf}
\alias{z_weight_for_length_inf}
\title{Defunct Functions}
\usage{
p_length_for_age_inf(q, age, male)

q_length_for_age_inf(p, age, male)

z_length_for_age_inf(q, age, male)

p_weight_for_age_inf(q, age, male)

q_weight_for_age_inf(p, age, male)

z_weight_for_age_inf(q, age, male)

p_weight_for_length_inf(q, length, male)

q_weight_for_length_inf(p, length, male)

z_weight_for_length_inf(q, length, male)
}
\arguments{
\item{q, p}{quantile or perenctile}

\item{age}{age in months}

\item{male}{0 = female; 1 = male}

\item{length}{length and height are term for stature depending on age, in centimeters}
}
\description{
These functions are fully defunct, no longer implemented, and will not work.
}
