% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbp-package.R
\docType{package}
\name{pedbp-package}
\alias{pedbp}
\alias{pedbp-package}
\title{PEDBP}
\description{
Tools for pediatric blood pressure distributions and growth standards.
}
\section{Notation}{

Across the package, arguments named \code{p} or beginning with \code{p_}
denote probabilities on the 0 to 1 scale. Arguments or values named
\code{percentile} denote percentile points on the 0 to 100 scale.

For example, use \code{p_sbp = 0.85} for the 85th percentile as a probability
input, and use \code{height_percentile = 85} for the 85th height percentile
as percentile points.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/pedbp/}
  \item \url{http://www.peteredewitt.com/pedbp/}
  \item Report bugs at \url{https://github.com/dewittpe/pedbp/issues/}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{peter.dewitt@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

Other contributors:
\itemize{
  \item Blake Martin \email{blake.martin@childrenscolorado.org} (\href{https://orcid.org/0000-0001-5683-8310}{ORCID}) [contributor]
  \item David Albers \email{david.albers@cuanschutz.edu} (\href{https://orcid.org/0000-0002-5369-526X}{ORCID}) [contributor]
  \item Tell Bennett \email{tell.bennett@cuanschutz.edu} (\href{https://orcid.org/0000-0003-1483-4236}{ORCID}) [contributor]
}

}
\keyword{internal}
