% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{pediatric_vital_sign_distributions}
\alias{pediatric_vital_sign_distributions}
\alias{p_bmi_for_age}
\alias{q_bmi_for_age}
\alias{z_bmi_for_age}
\alias{p_head_circ_for_age}
\alias{q_head_circ_for_age}
\alias{z_head_circ_for_age}
\alias{p_length_for_age_inf}
\alias{q_length_for_age_inf}
\alias{z_length_for_age_inf}
\alias{p_stature_for_age}
\alias{q_stature_for_age}
\alias{z_stature_for_age}
\alias{p_weight_for_age_inf}
\alias{q_weight_for_age_inf}
\alias{z_weight_for_age_inf}
\alias{p_weight_for_age}
\alias{q_weight_for_age}
\alias{z_weight_for_age}
\alias{p_weight_for_length_inf}
\alias{q_weight_for_length_inf}
\alias{z_weight_for_length_inf}
\alias{p_weight_for_stature}
\alias{q_weight_for_stature}
\alias{z_weight_for_stature}
\title{Pediatric Vital Sign Distributions}
\usage{
p_bmi_for_age(q, age, male)

q_bmi_for_age(p, age, male)

z_bmi_for_age(q, age, male)

p_head_circ_for_age(q, age, male)

q_head_circ_for_age(p, age, male)

z_head_circ_for_age(q, age, male)

p_length_for_age_inf(q, age, male)

q_length_for_age_inf(p, age, male)

z_length_for_age_inf(q, age, male)

p_stature_for_age(q, age, male)

q_stature_for_age(p, age, male)

z_stature_for_age(q, age, male)

p_weight_for_age_inf(q, age, male)

q_weight_for_age_inf(p, age, male)

z_weight_for_age_inf(q, age, male)

p_weight_for_age(q, age, male)

q_weight_for_age(p, age, male)

z_weight_for_age(q, age, male)

p_weight_for_length_inf(q, length, male)

q_weight_for_length_inf(p, length, male)

z_weight_for_length_inf(q, length, male)

p_weight_for_stature(q, height, male)

q_weight_for_stature(p, height, male)

z_weight_for_stature(q, height, male)
}
\arguments{
\item{q}{a vector of quantites}

\item{age}{numeric age, in months, for one patient}

\item{male}{integer value, 1 = male, 0 = female, indicating sex of the
patient}

\item{p}{a vector of probabilities}

\item{length}{length, in cm, of the patient (age under 3 years)}

\item{height}{height, in cm, of the patient (age 2 - 20 years)}
}
\description{
Based on the data provided by the CDC, provide the distribution function,
quantile function, and a z-score function for one of eight vital signs by
another vital sign, e.g., weight for age.  Values are based on an LMS
approach.
}
\examples{

p_bmi_for_age(q = 16.455, age = 32, male = 1)
z_bmi_for_age(q = 16.455, age = 32, male = 1)
q_bmi_for_age(p = 0.58, age = 32, male = 1)

p_head_circ_for_age(q = 16.455, age = 32, male = 1)
z_head_circ_for_age(q = 16.455, age = 32, male = 1)
q_head_circ_for_age(p = 0.58, age = 32, male = 1)

p_length_for_age_inf(q = 16.455, age = 32, male = 1)
z_length_for_age_inf(q = 16.455, age = 32, male = 1)
q_length_for_age_inf(p = 0.58, age = 32, male = 1)

p_stature_for_age(q = 16.455, age = 32, male = 1)
z_stature_for_age(q = 16.455, age = 32, male = 1)
q_stature_for_age(p = 0.58, age = 32, male = 1)

p_weight_for_age(q = 4.455, age = 24.1, male = 1)
q_weight_for_age(p = 0.05, age = 24.1, male = 1)
z_weight_for_age(q = 4.455, age = 24.1, male = 1)

p_weight_for_age_inf(q = 4.455, age = 2.41, male = 1)
q_weight_for_age_inf(p = 0.05, age = 2.41, male = 1)
z_weight_for_age_inf(q = 4.455, age = 2.41, male = 1)

p_weight_for_length_inf(q = 4.455, length = 55, male = 1)
q_weight_for_length_inf(p = 0.05, length = 55, male = 1)
z_weight_for_length_inf(q = 4.455, length = 55, male = 1)

p_weight_for_stature(q = 4.455, height = 95, male = 1)
q_weight_for_stature(p = 0.05, height = 95, male = 1)
z_weight_for_stature(q = 4.455, height = 95, male = 1)

}
\references{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
