% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_for_stature.R
\name{p_weight_for_height}
\alias{p_weight_for_height}
\alias{q_weight_for_height}
\alias{z_weight_for_height}
\alias{p_weight_for_length}
\alias{q_weight_for_length}
\alias{z_weight_for_length}
\title{Weight for Stature - Pediatric Growth Standard}
\usage{
p_weight_for_height(
  q,
  male,
  height,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

q_weight_for_height(
  p,
  male,
  height,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

z_weight_for_height(
  q,
  male,
  height,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

p_weight_for_length(
  q,
  male,
  length,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

q_weight_for_length(
  p,
  male,
  length,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

z_weight_for_length(
  q,
  male,
  length,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)
}
\arguments{
\item{q}{a vector of quantiles}

\item{male}{integer value, 1 = male, 0 = female}

\item{source}{a character string denoting the data source providing the
parameters needed for the estimate.}

\item{...}{pass through}

\item{p}{a vector of probabilities}

\item{age}{numeric age, in months}
}
\value{
The \code{p_} method return values from the estimated distribution
function.  \code{q_} methods return values from the estimated quantile
function.  \code{z_} methods return standard scores, equivalent to
\code{\link[stats]{qnorm}}.
}
\description{
Weight for stature quantile, distribution, and zscore function based on LMS data
from the CDC and WHO.
}
\details{
Length or height values are used.  Length is assess when the patient is lying
down versus height when the patient is standing.  There is an implication of
younger patients being in the _for_length set.  There is some overlap in
numeric values of length and height.
}
\references{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm},
\url{https://www.who.int/tools/child-growth-standards/standards}
}
