% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{zlms}
\alias{zlms}
\title{Distribution, Quantile, and Zscores by LMS values}
\usage{
zlms(x, l, m, s, ...)
}
\arguments{
\item{x}{quantile or percentile value}

\item{l, m, s}{the lms values}

\item{...}{pass through}
}
\description{
Non-exported functions ...
}
