% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bp_chart}
\alias{bp_chart}
\title{Blood Pressure Chart}
\usage{
bp_chart(age, height, height_unit = "percentile", ...)
}
\arguments{
\item{age}{age of the subject in years between 1 and 17.}

\item{height}{a numeric value}

\item{height_unit}{a character indicating if the \code{height} value is a
percentile (default), in inches, or centimeters.}

\item{...}{Pass through}
}
\description{
Produce a graphic showing the regions of blood pressure percentiles by age
(in years) and height (percentile).
}
