% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppPGSF}
\alias{cppPGSF}
\title{LMS data and distribution estimates}
\usage{
cppPGSF(metric, source, male, x, qp, type)
}
\arguments{
\item{age}{(in months)}

\item{stature}{(in cm)}
}
\description{
LMS data and distribution estimates
}
