% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{distribution-quantile-zscores-by-lms}
\alias{distribution-quantile-zscores-by-lms}
\alias{zlms}
\alias{plms}
\alias{qlms}
\title{Distribution, Quantile, and Z-scores by LMS values}
\usage{
zlms(x, l, m, s, ...)

plms(x, l, m, s, ...)

qlms(x, l, m, s, ...)
}
\arguments{
\item{x}{quantile or percentile value}

\item{l, m, s}{the lms values}

\item{...}{pass through}
}
\description{
Functions for getting estimated distribution, quantile, and standard scores
(z-scores) given LMS parameters.
}
\details{
L is the power in the Box-Cox transformation, M the median, and S a
generalized coefficient of variation. For a given standard score (z-score),
Z, the value X of interest is

\deqn{ X = \begin{cases} M (1 + LSZ)^{1/L} & L \neq 0 \\ M \exp(SZ) & L = 0.
\end{cases} }{M (1 + LSZ)^(1/L) for L != 0; M exp(SZ) for L = 0.}

To get the z-score for a value X:

\deqn{Z = \begin{cases} \frac{ \left(\frac{X}{M}\right)^{L} - 1 }{LS} & L
\neq 0 \\ \frac{\log\left(\frac{X}{M}}{S} & L = 0.}{ ( (X/M)^L - 1) / (LS)
for L != 0; log(X/M) / S for L = 0.}
}
\references{
Cole, Timothy J., and Pamela J. Green. "Smoothing reference
centile curves: the LMS method and penalized likelihood." Statistics in
medicine 11.10 (1992): 1305-1319.
}
