% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{get_lms}
\alias{get_lms}
\title{LMS Look Up Functions}
\usage{
get_lms(
  metric = c("bmi_for_age", "stature_for_age", "weight_for_age", "weight_for_stature",
    "head_circumference_for_age"),
  male,
  source = c("WHO", "CDC-2000"),
  age = NA_real_,
  stature = NA_real_
)
}
\arguments{
\item{metric}{section of the look up table}

\item{male}{0 = female; 1 = male}

\item{source}{"CDC-2000", or "WHO"}

\item{age}{in months}

\item{stature}{height/length}
}
\description{
Functions to look up LMS values from the internal data set lms_data
}
