% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms.R
\name{get_lms}
\alias{get_lms}
\title{LMS Look Up Functions}
\usage{
get_lms(
  metric = c("bmi_for_age", "stature_for_age", "weight_for_age", "weight_for_stature",
    "head_circumference_for_age"),
  male,
  source = c("CDC-WHO", "WHO", "CDC"),
  age = NA_real_,
  stature = NA_real_
)
}
\arguments{
\item{metric}{section of the look up tabl}

\item{male}{0 = female; 1 = male}

\item{source}{See Details in \code{\link{pediatric_growth_standards}}}

\item{age}{in months}

\item{stature}{height/length in centimeters}
}
\description{
Non-exported functions to look up LMS values from the internal data set
lms_data
}
\references{
\url{https://www.cdc.gov/growthcharts/who_charts.htm}
}
