% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_circumference_for_age.R
\name{head_circumference_for_age}
\alias{head_circumference_for_age}
\alias{p_head_circ_for_age}
\alias{q_head_circ_for_age}
\alias{z_head_circ_for_age}
\title{Head Circumference for Age - Pediatric Growth Standard}
\usage{
p_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

q_head_circ_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)

z_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgs_source", "CDC"),
  ...
)
}
\arguments{
\item{q}{a vector of quantiles}

\item{male}{integer value, 1 = male, 0 = female}

\item{age}{numeric age, in months}

\item{source}{a character string denoting the data source providing the
parameters needed for the estimate.}

\item{...}{pass through}

\item{p}{a vector of probabilities}
}
\value{
The \code{p_} method return values from the estimated distribution
function.  \code{q_} methods return values from the estimated quantile
function.  \code{z_} methods return standard scores, equivalent to
\code{\link[stats]{qnorm}}.
}
\description{
BMI for age quantile, distribution, and zscore function based on LMS data
from the CDC and WHO.
}
\details{
Note: CDC Recommends using WHO growth charts for infants and children ages 0
to 2 years of age in the U.S. and CDC growth charts to monitor growth for
children age 2 years and older in the U.S.
}
\references{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm},
\url{https://www.who.int/tools/child-growth-standards/standards}
}
