% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbp-defunt.R
\name{p_length_for_age_inf}
\alias{p_length_for_age_inf}
\alias{q_length_for_age_inf}
\alias{z_length_for_age_inf}
\alias{p_weight_for_age_inf}
\alias{q_weight_for_age_inf}
\alias{z_weight_for_age_inf}
\alias{p_weight_for_length_inf}
\alias{q_weight_for_length_inf}
\alias{z_weight_for_length_inf}
\title{Defunct Functions}
\usage{
p_length_for_age_inf(q, age, male)

q_length_for_age_inf(p, age, male)

z_length_for_age_inf(q, age, male)

p_weight_for_age_inf(q, age, male)

q_weight_for_age_inf(p, age, male)

z_weight_for_age_inf(q, age, male)

p_weight_for_length_inf(q, length, male)

q_weight_for_length_inf(p, length, male)

z_weight_for_length_inf(q, length, male)
}
\description{
Defunct Functions
}
