% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedbp-package.R
\docType{package}
\name{pedbp-package}
\alias{pedbp}
\alias{pedbp-package}
\title{PEDBP}
\description{
Data and utilities for estimating pediatric blood pressure percentiles by sex, age, and optionally height (stature) as described in Martin et.al. (2022) \doi{10.1001/jamanetworkopen.2022.36918}. Blood pressure percentiles for children under one year of age come from Gemelli et.al. (1990) \doi{10.1007/BF02171556}. Estimates of blood pressure percentiles for children at least one year of age are informed by data from the National Heart, Lung, and Blood Institute (NHLBI) and the Centers for Disease Control and Prevention (CDC) \doi{10.1542/peds.2009-2107C} or from Lo et.al. (2013) \doi{10.1542/peds.2012-1292}. The flowchart for selecting the informing data source comes from Martin et.al. (2022) \doi{10.1542/hpeds.2021-005998}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dewittpe/pedbp/}
}

}
\author{
\strong{Maintainer}: Peter DeWitt \email{peter.dewitt@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6391-0795}{ORCID})

Other contributors:
\itemize{
  \item Blake Martin \email{blake.martin@childrenscolorado.org} (\href{https://orcid.org/0000-0001-5683-8310}{ORCID}) [contributor]
  \item David Albers \email{david.albers@cuanschutz.edu} (\href{https://orcid.org/0000-0002-5369-526X}{ORCID}) [contributor]
  \item Tell Bennett \email{tell.bennett@cuanschutz.edu} (\href{https://orcid.org/0000-0003-1483-4236}{ORCID}) [contributor]
}

}
\keyword{internal}
