% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-standards.R
\name{pediatric_growth_standards}
\alias{pediatric_growth_standards}
\alias{p_bmi_for_age}
\alias{q_bmi_for_age}
\alias{z_bmi_for_age}
\alias{p_head_circ_for_age}
\alias{q_head_circ_for_age}
\alias{z_head_circ_for_age}
\alias{p_stature_for_age}
\alias{q_stature_for_age}
\alias{z_stature_for_age}
\alias{p_weight_for_age}
\alias{q_weight_for_age}
\alias{z_weight_for_age}
\alias{p_weight_for_stature}
\alias{q_weight_for_stature}
\alias{z_weight_for_stature}
\title{Pediatric Growth Standards}
\usage{
p_bmi_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_bmi_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_bmi_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_head_circ_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_stature_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_stature_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_stature_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_weight_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_weight_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_weight_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_weight_for_stature(
  q,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_weight_for_stature(
  p,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_weight_for_stature(
  q,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)
}
\arguments{
\item{q}{a vector of quantities}

\item{male}{integer value, 1 = male, 0 = female}

\item{age}{numeric age, in months}

\item{source}{a character string denoting the data source providing the
parameters needed for the estimate.  See Details.}

\item{...}{pass through}

\item{p}{a vector of probabilities}

\item{stature}{(height or length) in centimeters}
}
\value{
The \code{p_} method return values from the estimated distribution
function.  \code{q_} methods return values from the estimated quantile
function.  \code{z_} methods return standard scores, equivalent to
\code{\link[stats]{qnorm}}.
}
\description{
Based on the data provided by the CDC and the World Height Organization,
the distribution function, quantile function, and a z-score function for
several growth charts.
}
\details{
The \code{source} argument controls which data source is queried for the LMS
values.

Note: CDC Recommends using WHO growth charts for infants and children ages 0
to 2 years of age in the U.S. and CDC growth charts to monitor growth for
children age 2 years and older in the U.S.

To implement the CDC recommendation, the source \dQuote{CDC-WHO} will use
CDC for ages over 2, and WHO for under 2.  \dQuote{CDC} will use only the CDC
data, and \dQuote{WHO} will use only the WHO data.

Further, when \code{source = "CDC-WHO"}, use either source if only one
of the two is available.
}
\references{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm},
\url{https://www.who.int/tools/child-growth-standards/standards}
}
