% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-standards.R
\name{pediatric_growth_standards}
\alias{pediatric_growth_standards}
\alias{p_bmi_for_age}
\alias{q_bmi_for_age}
\alias{z_bmi_for_age}
\alias{p_head_circ_for_age}
\alias{q_head_circ_for_age}
\alias{z_head_circ_for_age}
\alias{p_stature_for_age}
\alias{q_stature_for_age}
\alias{z_stature_for_age}
\alias{p_weight_for_age}
\alias{q_weight_for_age}
\alias{z_weight_for_age}
\alias{p_weight_for_stature}
\alias{q_weight_for_stature}
\alias{z_weight_for_stature}
\title{Pediatric Growth Standards}
\usage{
p_bmi_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_bmi_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_bmi_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_head_circ_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_head_circ_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_stature_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_stature_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_stature_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_weight_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_weight_for_age(
  p,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_weight_for_age(
  q,
  male,
  age,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

p_weight_for_stature(
  q,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

q_weight_for_stature(
  p,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)

z_weight_for_stature(
  q,
  male,
  stature,
  source = getOption("pedbp_pgsf_source", "CDC-WHO"),
  ...
)
}
\arguments{
\item{q}{a vector of quantities}

\item{male}{integer value, 1 = male, 0 = female}

\item{age}{numeric age, in months}

\item{p}{a vector of probabilities}

\item{stature}{(height or length) in centimeters}
}
\value{
The \code{p_} method return values from the estimated distribution
function.  \code{q_} methods return values from the estimated quantile
function.  \code{z_} methods return standard scores, equivalent to
\code{\link[stats]{qnorm}}.
}
\description{
Based on the data provided by the CDC and the World Height Organization,
the distribution function, quantile function, and a z-score function for
several growth charts.
}
\examples{

#############################################################################
# BMI for Age

# A BMI of 18.2 for a 18.1 year old female is in the
p_bmi_for_age(q = 18.2, age = 18.1 * 12, male = 0)
# percentile.

# The z-score is the same as qnorm(p)
qnorm(p_bmi_for_age(q = 18.2, age = 18.1 * 12, male = 0))
z_bmi_for_age(q = 18.2, age = 18.1 * 12, male = 0)

# The 70th percentile of BMI for 15.4 year old males is
q_bmi_for_age(p = 0.70, age = 15.4 * 12, male = 1)

#############################################################################
# Stature (aka: Lenght or Height) for Age
bmi_for_age
head_circumference_for_age
stature_for_age
weight_for_age
weight_for_stature

#############################################################################
# Multiple patients, the age and male, length, height arguments can also be
# vectors
p_length_for_age_inf(q = 87, age = 28,  male = 0)
p_length_for_age_inf(q = 90, age = 30,  male = 1)
p_length_for_age_inf(q = c(87,90), age = c(28, 30),  male = c(0,1))


}
\references{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm},
\url{https://www.who.int/tools/child-growth-standards/standards}
}
